requirements.txt
contourpy==1.3.2
cupy-cuda12x==13.5.1
cycler==0.12.1
fastrlock==0.8.3
fonttools==4.58.5
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cufft-cu12==11.4.1.4
nvidia-nvjitlink-cu12==12.9.86
packaging==25.0
pillow==11.3.0
pyadi-iio==0.0.19
pylibiio==0.25
pyparsing==3.2.3
python-dateutil==2.9.0.post0
six==1.17.0
